refactor: update icon imports and enhance user avatar component for consistency
