+++ components/signin-google.tsx
@@ -4,7 +4,7 @@ import { signIn } from "next-auth/react";
 import { useState, useEffect } from "react";
 
 import { Button } from "@/components/ui/button";
-import { Icons } from "@/components/icons";
+import { Icons } from "@/components/ui/icons";
 
 export default function SignInWithGoogle() {
   const [isGoogleLoading, setIsGoogleLoading] = useState(false);
