+++ components/user-avatar.tsx
@@ -1,11 +1,11 @@
-import { AvatarProps } from "@radix-ui/react-avatar"
+import { AvatarProps } from "@radix-ui/react-avatar";
 
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
-import { Icons } from "@/components/icons"
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { Icons } from "@/components/ui/icons";
 
 interface UserAvatarProps extends AvatarProps {
-    user: { name: string; image: string, email: string };
-  }
+  user: { name: string; image: string; email: string };
+}
 
 export function UserAvatar({ user, ...props }: UserAvatarProps) {
   return (
@@ -19,5 +19,5 @@ export function UserAvatar({ user, ...props }: UserAvatarProps) {
         </AvatarFallback>
       )}
     </Avatar>
-  )
-} 
+  );
+}
